Fix leverage validation and improve endpoint status reporting
Adjusted leverage data type to float in `routes/alerts.py` to accommodate decimal values from AI Precision Calculator. Added more comprehensive endpoint status checks to `attached_assets/Pasted-appuser-srv-d4e9gh24d50c73dqs1t0-669f8644f8-mh26d-app-BASE-https-algogpt-trading-vm-onrender-co-1763624881417_1763624881421.txt` to better diagnose connectivity issues. Also added a new prediction entry for `1MBABYDOGEUSDT` to `data/ai_tracking/predictions.jsonl`.

+++ routes/alerts.py
@@ -260,7 +260,7 @@ class IngestReq(BaseModel):
     # אחת מהאפשרויות: qty או budget_usd (+ leverage)
     qty: Optional[float] = None
     budget_usd: Optional[float] = None
-    leverage: Optional[int] = None
+    leverage: Optional[float] = None  # Float to support AI Precision Calculator (e.g. 3.87x)
     # מידע נוסף (לא חובה)
     ticket_id: Optional[str] = None
     timeframe: Optional[str] = "15m"
